+++ internal/home/middlewares.go
@@ -30,7 +30,17 @@ const RequestBodySizeLimit = 64 * 1024
 func limitRequestBody(h http.Handler) (limited http.Handler) {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		var err error
-		r.Body, err = aghio.LimitReadCloser(r.Body, RequestBodySizeLimit)
+
+		var bodySizeLimit int64 = RequestBodySizeLimit
+		if u := r.URL; u.Path == "/control/access/set" {
+			// An exception for a poorly designed API.  Remove once
+			// the new, better API is up.
+			//
+			// See https://github.com/AdguardTeam/AdGuardHome/issues/2666.
+			bodySizeLimit *= 4
+		}
+
+		r.Body, err = aghio.LimitReadCloser(r.Body, bodySizeLimit)
 		if err != nil {
 			log.Error("limitRequestBody: %s", err)