+++ CHANGELOG.md
@@ -17,13 +17,20 @@ and this project adheres to
 ## [v0.105.1] - 2021-02-24
 -->
 
+### Changed
+
+- Increase the HTTP API request body size limit for the `/control/access/set`
+  API ([#2666]).
+
 ### Fixed
 
+- Set the request body size limit for HTTPS reqeusts as well.
 - Incorrect version tag in the Docker release ([#2663]).
 - DNSCrypt queries weren't marked as such in logs ([#2662]).
 
 [#2662]: https://github.com/AdguardTeam/AdGuardHome/issues/2662
 [#2663]: https://github.com/AdguardTeam/AdGuardHome/issues/2663
+[#2666]: https://github.com/AdguardTeam/AdGuardHome/issues/2666